tests/wpt/web-platform-tests/html/semantics/forms/the-selectmenu-element/selectmenu-required-attribute.tentative.html

<!DOCTYPE html>
<html lang="en">
<title>HTMLSelectMenuElement Test: required attribute</title>
<link rel="author" title="Ionel Popescu" href="mailto:iopopesc@microsoft.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
  selectmenu:required {
    border: 3px dashed rgb(255, 0, 0);
  }

  selectmenu:optional {
    border: 1px solid rgb(128, 128, 128);
  }
</style>

<selectmenu id="selectmenu0" required>
  <option>one</option>
  <option>two</option>
  <option>three</option>
</selectmenu>

<selectmenu id="selectmenu1">
  <option>one</option>
  <option>two</option>
  <option>three</option>
</selectmenu>

<selectmenu id="selectmenu2">
  <option>one</option>
  <option>two</option>
  <option>three</option>
</selectmenu>

<script>
function checkRequired(style) {
  assert_equals(style.borderWidth, '3px');
  assert_equals(style.borderStyle, 'dashed');
  assert_equals(style.borderColor, 'rgb(255, 0, 0)');
}

function checkOptional(style) {
  assert_equals(style.borderWidth, '1px');
  assert_equals(style.borderStyle, 'solid');
  assert_equals(style.borderColor, 'rgb(128, 128, 128)');
}

test(() => {
  const selectMenu0 = document.getElementById("selectmenu0");
  const selectMenu1 = document.getElementById("selectmenu1");
  const selectMenu2 = document.getElementById("selectmenu2");

  checkRequired(window.getComputedStyle(selectMenu0));
  checkOptional(window.getComputedStyle(selectMenu1));
  checkOptional(window.getComputedStyle(selectMenu2));
  selectMenu2.required = true;
  checkRequired(window.getComputedStyle(selectMenu2));
}, "Test required attribute");

</script>
