LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-selectmenu-element/selectmenu-required-attribute.tentative-expected.txt


FAIL Test required attribute assert_equals: expected "3px" but got "0px"
